Finishing touches, dict refactor
